+++ CHANGELOG.md
@@ -7,13 +7,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
-### Updated
+### Changed
 
 - `externalId` is generated from the API
 
 ### Removed
 
-- **[BREAKING]** `generateExternalId` from `form-service`
+- **[BREAKING]** `formService.generateExternalId`
 
 ## [16.1.0] - 2024-09-30